backend/requirements.txt
orjson
uvicorn[standard]
gunicorn
fastapi
fastapi-pagination
passlib[bcrypt]
python-jose[cryptography]
python-multipart

asyncpg
sqlalchemy
alembic
sqlmodel

# pydantic[email,dotenv]
# pydantic[email]
pydantic
email-validator

# loguru
apscheduler

pillow
# pandas

pytest
pytest-asyncio
httpx
aiosqlite

# https://fastapi.tiangolo.com/#optional-dependencies
